Implement the new interface in AE2 for network bytes information. (#52)
* Implements the new interface

* Delete commenteds

+++ src/main/java/thaumicenergistics/common/inventory/HandlerItemEssentiaCellCreative.java
@@ -111,7 +111,7 @@ public class HandlerItemEssentiaCellCreative extends HandlerItemEssentiaCell {
      * Used types based on partition table.
      */
     @Override
-    public int getUsedTypes() {
+    public long getUsedTypes() {
         return this.partitionAspects.size();
     }
